move STA_configuration.py to  fastvideo/attention/backends add diagram to sta docs
